requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.8
django-crispy-forms==2.3
gunicorn==22.0.0
packaging==24.1
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==5.3.0
